requirements.txt
asgiref==3.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-extensions==3.1.5
django-3-jet==1.0.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Pillow==9.0.1
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
